:brick:[#28] location collectionview datasource 분리
